+++ main/model.h
@@ -12,8 +12,8 @@ class ModelInterface {
 public:
   virtual ~ModelInterface(){};
   virtual std::vector<objects::TriangleObj *> triangles() const = 0;
-  virtual double height() const = 0;
-  virtual double sideSize() const = 0;
+  virtual float height() const = 0;
+  virtual float sideSize() const = 0;
   // returns true if there is no object assigned to a model.
   virtual bool empty() const = 0;
 };
@@ -22,18 +22,18 @@ class Model : public ModelInterface {
 public:
   std::vector<objects::TriangleObj *> triangles() const;
 
-  double height() const { return height_; }
-  double sideSize() const { return sideSize_; }
+  float height() const { return height_; }
+  float sideSize() const { return sideSize_; }
   bool empty() const override {
-    return std::max<double>(height(), sideSize()) <= constants::kAccuracy;
+    return std::max<float>(height(), sideSize()) <= constants::kAccuracy;
   }
 
   static std::unique_ptr<Model> NewLoadFromObjectFile(std::string_view path);
-  static std::unique_ptr<Model> NewReferenceModel(double sideSize);
+  static std::unique_ptr<Model> NewReferenceModel(float sideSize);
 
 private:
   std::vector<std::unique_ptr<objects::TriangleObj>> triangles_;
-  double height_, sideSize_;
+  float height_, sideSize_;
 };
 
 #endif
